feat(cli): no more spek
